+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 
 ## [Unreleased]
+
+## [0.0.2] — 2019-09-09
 ### Added
 - Factory function for creating mappers.
 - Macro for defining mappers in a namespace. 
@@ -25,4 +27,5 @@ and this project adheres to
 Released without _CHANGELOG.md_.
 
 [0.0.1]: https://github.com/b-social/jason/compare/0.0.1...0.0.1
-[Unreleased]: https://github.com/b-social/jason/compare/0.0.1...HEAD
+[0.0.2]: https://github.com/b-social/jason/compare/0.0.1...0.0.2
+[Unreleased]: https://github.com/b-social/jason/compare/0.0.2...HEAD